+++ openedx_tagging/core/tagging/models/__init__.py
@@ -9,3 +9,4 @@ from .system_defined import (
     UserSystemDefinedTaxonomy,
     LanguageTaxonomy,
 )
+from .import_export import TagImportTask
